requirements.txt
bottle==0.12.13
paste==2.0.3
Pillow==4.2.1
scandir==1.5
imageio==2.2.0
